Imported Upstream version 0.6.58

+++ src/extern/pygments/formatters/html.py
@@ -5,33 +5,33 @@
 
     Formatter for HTML output.
 
-    :copyright: 2006-2008 by Georg Brandl, Armin Ronacher.
-    :license: BSD, see LICENSE for more details.
+    :copyright: Copyright 2006-2010 by the Pygments team, see AUTHORS.
+    :license: BSD, see LICENSE for details.
 """
-import sys, os
-import StringIO
 
-try:
-    set
-except NameError:
-    from sets import Set as set
+import os
+import sys
+import StringIO
 
 from pygments.formatter import Formatter
 from pygments.token import Token, Text, STANDARD_TYPES
-from pygments.util import get_bool_opt, get_int_opt, get_list_opt
+from pygments.util import get_bool_opt, get_int_opt, get_list_opt, bytes
 
 
 __all__ = ['HtmlFormatter']
 
 
-def escape_html(text):
-    """Escape &, <, > as well as single and double quotes for HTML."""
-    return text.replace('&', '&amp;').  \
-                replace('<', '&lt;').   \
-                replace('>', '&gt;').   \
-                replace('"', '&quot;'). \
-                replace("'", '&#39;')
+_escape_html_table = {
+    ord('&'): u'&amp;',
+    ord('<'): u'&lt;',
+    ord('>'): u'&gt;',
+    ord('"'): u'&quot;',
+    ord("'"): u'&#39;',
+}
 
+def escape_html(text, table=_escape_html_table):
+    """Escape &, <, > as well as single and double quotes for HTML."""
+    return text.translate(table)
 
 def get_random_id():
     """Return a random id for javascript fields."""
@@ -186,7 +186,9 @@ class HtmlFormatter(Formatter):
 
     `style`
         The style to use, can be a string or a Style subclass (default:
-        ``'default'``).
+        ``'default'``). This option has no effect if the `cssfile`
+        and `noclobber_cssfile` option are given and the file specified in
+        `cssfile` exists.
 
     `noclasses`
         If set to true, token ``<span>`` tags will not use CSS classes, but
@@ -222,6 +224,12 @@ class HtmlFormatter(Formatter):
         path, the file's path will be assumed to be relative to the main output
         file's path, if the latter can be found. The stylesheet is then written
         to this file instead of the HTML file. *New in Pygments 0.6.*
+
+    `noclobber_cssfile`
+        If `cssfile` is given and the specified file exists, the css file will
+        not be overwritten. This allows the use of the `full` option in
+        combination with a user specified css file. Default is ``False``.
+        *New in Pygments 1.1.*
 
     `linenos`
         If set to ``'table'``, output line numbers as a table with two cells,
@@ -273,6 +281,10 @@ class HtmlFormatter(Formatter):
         If set to a nonempty string, e.g. ``foo``, the formatter will wrap each
         output line in an anchor tag with a ``name`` of ``foo-linenumber``.
         This allows easy linking to certain lines. *New in Pygments 0.9.*
+
+    `anchorlinenos`
+        If set to `True`, will wrap line numbers in <a> tags. Used in
+        combination with `linenos` and `lineanchors`.
 
 
     **Subclassing the HTML formatter**
@@ -330,14 +342,16 @@ class HtmlFormatter(Formatter):
 
     def __init__(self, **options):
         Formatter.__init__(self, **options)
-        self.title = self._encodeifneeded(self.title)
+        self.title = self._decodeifneeded(self.title)
         self.nowrap = get_bool_opt(options, 'nowrap', False)
         self.noclasses = get_bool_opt(options, 'noclasses', False)
         self.classprefix = options.get('classprefix', '')
-        self.cssclass = self._encodeifneeded(options.get('cssclass', 'highlight'))
-        self.cssstyles = self._encodeifneeded(options.get('cssstyles', ''))
-        self.prestyles = self._encodeifneeded(options.get('prestyles', ''))
-        self.cssfile = self._encodeifneeded(options.get('cssfile', ''))
+        self.cssclass = self._decodeifneeded(options.get('cssclass', 'highlight'))
+        self.cssstyles = self._decodeifneeded(options.get('cssstyles', ''))
+        self.prestyles = self._decodeifneeded(options.get('prestyles', ''))
+        self.cssfile = self._decodeifneeded(options.get('cssfile', ''))
+        self.noclobber_cssfile = get_bool_opt(options, 'noclobber_cssfile', False)
+
         linenos = options.get('linenos', False)
         if linenos == 'inline':
             self.linenos = 2
@@ -352,6 +366,7 @@ class HtmlFormatter(Formatter):
         self.nobackground = get_bool_opt(options, 'nobackground', False)
         self.lineseparator = options.get('lineseparator', '\n')
         self.lineanchors = options.get('lineanchors', '')
+        self.anchorlinenos = options.get('anchorlinenos', False)
         self.hl_lines = set()
         for lineno in get_list_opt(options, 'hl_lines', []):
             try:
@@ -359,22 +374,21 @@ class HtmlFormatter(Formatter):
             except ValueError:
                 pass
 
-        self._class_cache = {}
         self._create_stylesheet()
 
     def _get_css_class(self, ttype):
         """Return the css class of this token type prefixed with
         the classprefix option."""
-        if ttype in self._class_cache:
-            return self._class_cache[ttype]
-        return self.classprefix + _get_ttype_class(ttype)
+        ttypeclass = _get_ttype_class(ttype)
+        if ttypeclass:
+            return self.classprefix + ttypeclass
+        return ''
 
     def _create_stylesheet(self):
         t2c = self.ttype2class = {Token: ''}
         c2s = self.class2style = {}
-        cp = self.classprefix
         for ttype, ndef in self.style:
-            name = cp + _get_ttype_class(ttype)
+            name = self._get_css_class(ttype)
             style = ''
             if ndef['color']:
                 style += 'color: #%s; ' % ndef['color']
@@ -433,10 +447,12 @@ class HtmlFormatter(Formatter):
                          (prefix(''), self.style.highlight_color))
         return '\n'.join(lines)
 
-    def _encodeifneeded(self, value):
-        if not self.encoding or isinstance(value, str):
-            return value
-        return value.encode(self.encoding)
+    def _decodeifneeded(self, value):
+        if isinstance(value, bytes):
+            if self.encoding:
+                return value.decode(self.encoding)
+            return value.decode()
+        return value
 
     def _wrap_full(self, inner, outfile):
         if self.cssfile:
@@ -449,17 +465,19 @@ class HtmlFormatter(Formatter):
                     if not filename or filename[0] == '<':
                         # pseudo files, e.g. name == '<fdopen>'
                         raise AttributeError
-                    cssfilename = os.path.join(os.path.dirname(filename), self.cssfile)
+                    cssfilename = os.path.join(os.path.dirname(filename),
+                                               self.cssfile)
                 except AttributeError:
                     print >>sys.stderr, 'Note: Cannot determine output file name, ' \
                           'using current directory as base for the CSS file name'
                     cssfilename = self.cssfile
-            # write CSS file
+            # write CSS file only if noclobber_cssfile isn't given as an option.
             try:
-                cf = open(cssfilename, "w")
-                cf.write(CSSFILE_TEMPLATE %
-                         {'styledefs': self.get_style_defs('body')})
-                cf.close()
+                if not os.path.exists(cssfilename) or not self.noclobber_cssfile:
+                    cf = open(cssfilename, "w")
+                    cf.write(CSSFILE_TEMPLATE %
+                            {'styledefs': self.get_style_defs('body')})
+                    cf.close()
             except IOError, err:
                 err.strerror = 'Error writing CSS file: ' + err.strerror
                 raise
@@ -490,19 +508,53 @@ class HtmlFormatter(Formatter):
         mw = len(str(lncount + fl - 1))
         sp = self.linenospecial
         st = self.linenostep
+        la = self.lineanchors
+        aln = self.anchorlinenos
+        nocls = self.noclasses
         if sp:
-            ls = '\n'.join([(i%st == 0 and
-                             (i%sp == 0 and '<span class="special">%*d</span>'
-                              or '%*d') % (mw, i)
-                             or '')
-                            for i in range(fl, fl + lncount)])
-        else:
-            ls = '\n'.join([(i%st == 0 and ('%*d' % (mw, i)) or '')
-                            for i in range(fl, fl + lncount)])
+            lines = []
 
-        yield 0, ('<table class="%stable">' % self.cssclass +
-                  '<tr><td class="linenos"><pre>' +
-                  ls + '</pre></td><td class="code">')
+            for i in range(fl, fl+lncount):
+                if i % st == 0:
+                    if i % sp == 0:
+                        if aln:
+                            lines.append('<a href="#%s-%d" class="special">%*d</a>' %
+                                         (la, i, mw, i))
+                        else:
+                            lines.append('<span class="special">%*d</span>' % (mw, i))
+                    else:
+                        if aln:
+                            lines.append('<a href="#%s-%d">%*d</a>' % (la, i, mw, i))
+                        else:
+                            lines.append('%*d' % (mw, i))
+                else:
+                    lines.append('')
+            ls = '\n'.join(lines)
+        else:
+            lines = []
+            for i in range(fl, fl+lncount):
+                if i % st == 0:
+                    if aln:
+                        lines.append('<a href="#%s-%d">%*d</a>' % (la, i, mw, i))
+                    else:
+                        lines.append('%*d' % (mw, i))
+                else:
+                    lines.append('')
+            ls = '\n'.join(lines)
+
+        # in case you wonder about the seemingly redundant <div> here: since the
+        # content in the other cell also is wrapped in a div, some browsers in
+        # some configurations seem to mess up the formatting...
+        if nocls:
+            yield 0, ('<table class="%stable">' % self.cssclass +
+                      '<tr><td><div class="linenodiv" '
+                      'style="background-color: #f0f0f0; padding-right: 10px">'
+                      '<pre style="line-height: 125%">' +
+                      ls + '</pre></div></td><td class="code">')
+        else:
+            yield 0, ('<table class="%stable">' % self.cssclass +
+                      '<tr><td class="linenos"><div class="linenodiv"><pre>' +
+                      ls + '</pre></div></td><td class="code">')
         yield 0, dummyoutfile.getvalue()
         yield 0, '</td></tr></table>'
 
@@ -514,10 +566,27 @@ class HtmlFormatter(Formatter):
         num = self.linenostart
         mw = len(str(len(lines) + num - 1))
 
-        if sp:
+        if self.noclasses:
+            if sp:
+                for t, line in lines:
+                    if num%sp == 0:
+                        style = 'background-color: #ffffc0; padding: 0 5px 0 5px'
+                    else:
+                        style = 'background-color: #f0f0f0; padding: 0 5px 0 5px'
+                    yield 1, '<span style="%s">%*s</span> ' % (
+                        style, mw, (num%st and ' ' or num)) + line
+                    num += 1
+            else:
+                for t, line in lines:
+                    yield 1, ('<span style="background-color: #f0f0f0; '
+                              'padding: 0 5px 0 5px">%*s</span> ' % (
+                              mw, (num%st and ' ' or num)) + line)
+                    num += 1
+        elif sp:
             for t, line in lines:
                 yield 1, '<span class="lineno%s">%*s</span> ' % (
-                    num%sp == 0 and ' special' or '', mw, (num%st and ' ' or num)) + line
+                    num%sp == 0 and ' special' or '', mw,
+                    (num%st and ' ' or num)) + line
                 num += 1
         else:
             for t, line in lines:
@@ -536,15 +605,29 @@ class HtmlFormatter(Formatter):
                 yield 0, line
 
     def _wrap_div(self, inner):
+        style = []
+        if (self.noclasses and not self.nobackground and
+            self.style.background_color is not None):
+            style.append('background: %s' % (self.style.background_color,))
+        if self.cssstyles:
+            style.append(self.cssstyles)
+        style = '; '.join(style)
+
         yield 0, ('<div' + (self.cssclass and ' class="%s"' % self.cssclass)
-                  + (self.cssstyles and ' style="%s"' % self.cssstyles) + '>')
+                  + (style and (' style="%s"' % style)) + '>')
         for tup in inner:
             yield tup
         yield 0, '</div>\n'
 
     def _wrap_pre(self, inner):
-        yield 0, ('<pre'
-                  + (self.prestyles and ' style="%s"' % self.prestyles) + '>')
+        style = []
+        if self.prestyles:
+            style.append(self.prestyles)
+        if self.noclasses:
+            style.append('line-height: 125%')
+        style = '; '.join(style)
+
+        yield 0, ('<pre' + (style and ' style="%s"' % style) + '>')
         for tup in inner:
             yield tup
         yield 0, '</pre>'
@@ -555,11 +638,11 @@ class HtmlFormatter(Formatter):
         Yield individual lines.
         """
         nocls = self.noclasses
-        enc = self.encoding
         lsep = self.lineseparator
         # for <span style=""> lookup only
         getcls = self.ttype2class.get
         c2s = self.class2style
+        escape_table = _escape_html_table
 
         lspan = ''
         line = ''
@@ -574,10 +657,7 @@ class HtmlFormatter(Formatter):
                 cls = self._get_css_class(ttype)
                 cspan = cls and '<span class="%s">' % cls or ''
 
-            if enc:
-                value = value.encode(enc)
-
-            parts = escape_html(value).split('\n')
+            parts = value.translate(escape_table).split('\n')
 
             # for all but the last line
             for part in parts[:-1]:
@@ -619,7 +699,14 @@ class HtmlFormatter(Formatter):
             if t != 1:
                 yield t, value
             if i + 1 in hls: # i + 1 because Python indexes start at 0
-                yield 1, '<span class="hll">%s</span>' % value
+                if self.noclasses:
+                    style = ''
+                    if self.style.highlight_color is not None:
+                        style = (' style="background-color: %s"' %
+                                 (self.style.highlight_color,))
+                    yield 1, '<span%s>%s</span>' % (style, value)
+                else:
+                    yield 1, '<span class="hll">%s</span>' % value
             else:
                 yield 1, value
 
@@ -631,7 +718,7 @@ class HtmlFormatter(Formatter):
         """
         return self._wrap_div(self._wrap_pre(source))
 
-    def format(self, tokensource, outfile):
+    def format_unencoded(self, tokensource, outfile):
         """
         The formatting process uses several nested generators; which of
         them are used is determined by the user's options.
